apinre/requirements.txt
Django==2.0
django-cors-headers==2.2.0
djangorestframework==3.8.2
Pillow==5.1.0
pytz==2018.4
